### 2017-01-11 #158 @whedon
[PRE REVIEW]: remBoot: An R package for Random Encounter Modelling
**Submitting author:** @arcaravaggi (<a href="http://orcid.org/0000-0002-1763-8970">Anthony Caravaggi</a>)
**Repository:** <a href="https://github.com/arcaravaggi/remBoot" target ="_blank">https://github.com/arcaravaggi/remBoot</a>
**Version:** 0.0.0.9000
