### 2017-01-11 #158 @karthik
@arfon Sure, np.
