### 2017-01-11 #158 @arcaravaggi
@karthik - Perhaps @jsta and/or @amoeba would be appropriate reviewers? Both list ecology as a topic of interest.
