### 2017-01-11 1773e72
Update paper.md
