### 2017-01-11 2e0b658
Update paper.md
