### 2017-01-11 ab6dc4d
Update paper.md
