### 2017-02-03 #158 @amoeba
Hey @arcaravaggi, sorry I didn't catch your first mention. Unless @jsta is eager to review this one, I can do this before Monday. I'm not familiar with Random Encounter Modeling but I can probably wrap my head around it.

At a first (quick) glance, I expect I'm going to have comments about:

- The readme looks to be missing sections
- The paper looks to have too much info but I'm not sure and it may be totally fine
- The package is missing automated tests but I'm not sure the package is complex to warrant much in the way of tests

A more thorough review to come when I get a chance this weekend.
