### 2017-02-03 #158 @arcaravaggi
Hi @amoeba, thanks for agreeing to review the package. I appreciate your time and feedback.
