### 2017-02-04 #176 @whedon
[REVIEW]: remBoot: An R package for Random Encounter Modelling
**Submitting author:** arcaravaggi (<a href="http://orcid.org/0000-0002-1763-8970">Anthony Caravaggi</a>)
**Repository:** <a href="https://github.com/arcaravaggi/remBoot" target ="_blank">https://github.com/arcaravaggi/remBoot</a>
**Version:** 0.0.0.9000
**Archive:** <a href="http://dx.doi.org/10.6084/m9.figshare.4684894" target="_blank">10.6084/m9.figshare.4684894</a>

- [x] **Version:** Does the release version given match the GitHub release (0.0.0.9000)?
- [x] **Authorship:** Has the submitting author (arcaravaggi) made major contributions to the software?
