### 2017-02-04 #158 @arfon
> Hey @arcaravaggi, sorry I didn't catch your first mention. Unless @jsta is eager to review this one, I can do this before Monday. I'm not familiar with Random Encounter Modeling but I can probably wrap my head around it.

Awesome, thanks @amoeba. Please head over to #176 for the actual review issue (with checklist etc).
