### 2017-02-04 #176 @amoeba
I've had a short break to sit down with this submission. I will finish up tomorrow at some point.

Overall, things look pretty good and I only have minor suggestions so far. I think packages like this that implement methods from the literature are great and think this would be a useful package for others.

I've copied a subset of the checklist below to make comments on particular items in the checklist.

- **Version: Does the release version given match the GitHub release (0.0.0.9000)?**

No. There is no release(s) on GitHub. Please create a release for 0.0.0.9000.

- **Authorship: Has the submitting author (arcaravaggi) made major contributions to the software?**

Possibly and more-than-likely. The majority of the code was committed to the GitHub repository in one commit so the provenance of the code itself can't be verified. I checked this off as OK anyway.

- **Installation: Does installation proceed as outlined in the documentation?**

There is no documentation for how to install this software. Please include a section in the base README providing a complete description of how to install the software.

- **A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?**

The statement given at the top of the README is sufficient but I suggest the reviewer add a sentence or two about why I want to use (1) this piece of software over others (are there others?) and (2) why I would want to use REM for my camera trap data. It looks like this copy is already in the included paper and vignette so this may be easy for the author to move into the README.

- **Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).**

The software includes one vignette which appears sufficient. I would encourage the author to (1) include a minimal worked example in the README and also fill in more details in the examples within function documentation (see my comment below).

- **Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g. API method documentation)?**

  - Function documentation is missing argument types. Suggest adding them. For example, in `remBoot`, change

> tm	= camera hours

to

> tm (numeric) The number of hours ...

  - From reading over the vignette, it appears the core functionality of the package is made up of two functions: `rem` and `remBoot`. However, a number of other functions are exported. If these are not user-facing functions, I suggest not.
  - Examples within function documentation is minimal. I suggest the reviewer spend some time expanding the examples in `rem` and `remBoot` so they are complete.

- **Automated tests: Are there automated tests or manual steps described so that the function of the software can be verified?**

No tests are present. Given the simplicity of this package, it shouldn't take much in the way of tests to provide -- at a minimum -- a small set of regression tests.

- **Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support**

No, please edit the README to include this information.

I'll move onto verifying the software tomorrow. Please reach out if you have any questions about my comments so far.
