### 2017-02-06 #176 @amoeba
I just had a chance to sit down to try to verify the functionality. Things seem in order but I have some comments about how REM is calculated and a lot of general comments about the code and help docs.

- **Functionality: Have the functional claims of the software been confirmed?**

  - The `rem` function appears to do what is outlined in the methods. I found that `r` and `theta` used in the calculation of REM are averaged which appears to be what is outlined in the methods:

    > Means of r and θ across all relevant detections were used in a separate REM for each 1 km square.

    If this is the correct way to implement the REM method then the software should work correctly given sane inputs. However -- as a note -- the above quote implies the means are taken within a site rather than for the study as a whole. I think if the user calculates their REM densities using `remBoot`, they will be fine because `remBoot` splits by site. But if they calculate their REM densities directly with the `rem` function, they may need to be careful. Does this sound right @arcaravaggi?

    Also, the means for `r` and `theta` are calculated with `na.rm =TRUE`. What do NAs in these two values mean? I didn't find a place in the methods section of the linked papers about handling NAs. Maybe I just didn't look carefully enough.

  - The `remBoot` function appears to do what is outlined.

  Please comment on these two issues so I can check this item off.

General code comments:

- The user-facing functions perform little sanity-checking on input arguments and it seems like bad input might cause a result to be calculated that is non-sense (with no errors or warnings thrown). The JOSS review doesn't get into this level of detail on the code but I encourage the author to consider this if possible.
- It takes a bit of work to find out what columns the input data.frame should have. Given that the two user-facing functions are `rem` and `remBoot`, it would be great if each of these help pages included that information.
- The `nboots` argument is provided in the function signature of `remBoot` but never used in the function body. It appears the user can't configure this.
- The example code for `split_dat` includes the body of the code itself. A more useful example should be included if the user is supposed to use it.
- The function `grpDat_func` should be called something more useful to the user if it's intended to be a user-facing function. It's body should also not be defined in the example section.
- The function `bsD` mentions a function `remsD` which does not exist in the package. Something isn't right here. I see that `remsD` is a variable defined in the function body of `remBoot`. What's going on here?
- Help docs on most pages could be improved greatly. I can give more specific details if the author needs.
- Help doc on `rem` links to web and not to internal vignette. Seems like it's the same vignette that is included in the package and the local vignette would be vastly preferred to an online link to it.

That's quite a long list so I apologize in advance. I think I've covered enough topics for you to get started. So please let me know if I can help by answering questions or elaborating. If any of my suggestions seem difficult to implement, I might be able to assist in addressing them in a meaningful way (writing code, writing docs, etc).
