### 2017-02-06 f26a9bc
Update paper.md
