+++ paper/paper.md
@@ -35,7 +35,7 @@ Caravaggi, A., Zaccaroni, M., Riga, F., Schai-Braun, S.C., Dick, J.T., Montgomer
 
 Manzo, E., Bartolommei, P., Rowcliffe, J.M. and Cozzolino, R. (2011). Estimation of population density of European pine marten in central Italy using camera trapping. Acta Theriol. 57, 165 – 172. [DOI: 10.1007/s13364-011-0055-8](http://link.springer.com/article/10.1007/s13364-011-0055-8)
 
-Oliveira-Santos, L.G.R., Zucco, C.A., Antunes, P.C. and Crawshaw, P.G. (2010). Is it possible to individually identify mammals with no natural markings using camera-traps? A controlled case-study with lowland tapirs. Mamm Biol. 75, 375 – 378. [DOI: 10.1016/j.mambio.2009.08.005](10.1016/j.mambio.2009.08.005)
+Oliveira-Santos, L.G.R., Zucco, C.A., Antunes, P.C. and Crawshaw, P.G. (2010). Is it possible to individually identify mammals with no natural markings using camera-traps? A controlled case-study with lowland tapirs. Mamm Biol. 75, 375 – 378. [DOI: 10.1016/j.mambio.2009.08.005](http://www.sciencedirect.com/science/article/pii/S161650470900130X)
 
 Rowcliffe, J.M., Field, J., Turvey, S.T. and Carbone, C. (2008). Estimating animal density using camera traps without the need for individual recognition. J Appl Ecol. 45, 1228 – 1236. [DOI: 10.1111/j.1365-2664.2008.01473.x] (http://onlinelibrary.wiley.com/doi/10.1111/j.1365-2664.2008.01473.x/abstract)
 
