### 2017-02-06 #176 @arcaravaggi
Thanks for the detailed feedback, @amoeba. I'll do my very best to look at each issue in detail this week and will respond as soon as I can.
