### 2017-02-09 57d67cd
Installation instructions
# Installation

To install remBoot from the GitHub repo, use the following commands:

install.packages("githubinstall")
library(githubinstall)
githubinstall("remBoot")
