### 2017-02-09 2a2e47f
Introduction text
Moved package introduction text from vignette to README
