### 2017-02-09 72c8913
Text moved
Introductory text moved to README
