### 2017-02-09 7885cce
rem corrections
Corrected function documentation, described data requirements and
expanded the example contained within.
