### 2017-02-09 29cf096
Contribution & support
Added development text - info for contributors and re: errors - to
README.
