### 2017-02-09 b6484cc
rem
Minor changes to example
