### 2017-02-09 ae4a25c
README example
Added example to README
