+++ vignettes/remBoot.Rmd
@@ -10,12 +10,6 @@ vignette: >
   %\VignetteEncoding{UTF-8}
 ---
 
-## remBoot Introduction
-
-remBoot is an implementation of the Random Encounter Model (REM) developed by Rowcliffe _et al._ (2008). The REM is based on brownian motion and allows the estimation of animal densities for a given survey area. The REM is particularly useful in that species do not need to be marked or exhibit individual markings for estimates to be calculated. This package contains a number of functions which allow the user to calculate densities for one or more sites and bootstrap their data to calculate variance. 
-
-## Example
-
 Read the packaged data file. These data are from a study of hares in Northern Ireland, published as Caravaggi _et al._ (2016). The first 5 columns of any data set must follow the same format in terms of column contents.
 
 ```{r include = TRUE}
@@ -29,12 +23,11 @@ Alternatively, generate relevant simulated data.
 ```{r include = TRUE}
 
 remDat <- data.frame(cbind(site <- sample(1:4, 200, replace=TRUE),
-                           cam <- sample(1:20, 200, replace = TRUE),
                            count <- sample(1:3, 200, replace = TRUE),
                            dist <- runif(200, min=0, max=0.006),
                            theta <- runif(200, min=0, max=0.6)))
 
-colnames(remDat) <- c("site","cam", "count", "dist", "theta")
+colnames(remDat) <- c("site", "count", "dist", "theta")
 
 ```
 
@@ -112,11 +105,12 @@ remsSD
 ```
 
 
-Alternatively, we can forgo all the above and use the __remBoot__ function to split and group data and calculate variance (SD and/or 95% confidence intervals) in one go. __remBoot__ requires that all sites have the same values for tm and v. As before, we'll define tm and v a priori:
+Alternatively, we can forgo all the above and use the __remBoot__ function to split and group data and calculate variance (SD and/or 95% confidence intervals) in one go. __remBoot__ requires that all sites have the same values for tm and v. As before, we'll define _tm_,  _v_ and _nboots_ a priori:
 
 ```{r include = TRUE}
 tm <- 1880
 v <- 0.89
+nboots <- 1000
 ```
 
 
