### 2017-02-09 42c1daa
Delete remBoot.knit.md
