+++ vignettes/remBoot.knit.md
@@ -1,312 +1,0 @@
----
-title: "remBoot User Guide"
-author: "Anthony Caravaggi (@thonoir; arcaravaggi.co.uk)"
-date: "2017-01-09"
-output: rmarkdown::html_vignette
-fig_caption: yes
-vignette: >
-  %\VignetteIndexEntry{Vignette Title}
-  %\VignetteEngine{knitr::rmarkdown}
-  %\VignetteEncoding{UTF-8}
----
-
-## remBoot Introduction
-
-remBoot is an implementation of the Random Encounter Model (REM) developed by Rowcliffe et al. (2008). The REM is based on brownian motion and allows the estimation of animal densities for a given survey area. The REM is particularly useful in that species do not need to be marked or exhibit individual markings for estimates to be calculated. This package contains a number of functions which allow the user to calculate densities for one or more sites and bootstrap their data to calculate variance. 
-
-## Example
-
-Read the provided data file. These data are from a study of hares in Northern Ireland, published as Caravaggi et al. (2016). The first 5 columns of any data set must follow the same format in terms of column contents.
-
-
-```r
-data(hDat)
-```
-
-Alternatively, generate relevant simulated data. 
-
-
-```r
-remDat <- data.frame(cbind(site <- sample(1:4, 200, replace=TRUE),
-                           cam <- sample(1:20, 200, replace = TRUE),
-                           count <- sample(1:3, 200, replace = TRUE),
-                           dist <- runif(200, min=0, max=0.006),
-                           theta <- runif(200, min=0, max=0.6)))
-
-colnames(remDat) <- c("site","cam", "count", "dist", "theta")
-```
-
-
-Now we'll check the first few rows of data, check the sample size for each site, and check data consistency.
-
-
-```r
-head(hDat) 
-```
-
-```
-##   site cam count  dist theta
-## 1    1   2     1 0.003  0.67
-## 2    1   2     1 0.003  0.67
-## 3    1   2     1 0.003  0.67
-## 4    1   3     1 0.004  0.39
-## 5    1   3     1 0.003  0.18
-## 6    1   3     1 0.005  0.10
-```
-
-```r
-tapply(hDat[,4], hDat[,1], length) 
-```
-
-```
-##  1  2  3  4  5 
-## 36 16 20 19 27
-```
-
-```r
-sum(tapply(hDat[,4], hDat[,1], length)) == nrow(hDat) 
-```
-
-```
-## [1] TRUE
-```
-
-We'll use two functions to split the data according to survey site (__res_dat__) and store the number of rows for each site (__split_dat__). These data will be used for iterative resampling during bootstrapping. The first step MUST be called grpSize as it is used in the core REM function.
-
-
-```r
-grpSize <- res_dat(hDat)
-
-grpDat <- split_dat(hDat)
-```
-
-Check that the code has done what we wanted
-
-
-```r
-lapply(grpDat, head)
-```
-
-```
-## [[1]]
-##   site cam count  dist theta
-## 1    1   2     1 0.003  0.67
-## 2    1   2     1 0.003  0.67
-## 3    1   2     1 0.003  0.67
-## 4    1   3     1 0.004  0.39
-## 5    1   3     1 0.003  0.18
-## 6    1   3     1 0.005  0.10
-## 
-## [[2]]
-##    site cam count  dist theta
-## 37    2   2     1 0.001  1.37
-## 38    2   2     1 0.006  0.06
[…]
