### 2017-02-09 166eefe
Update vignette files
