+++ vignettes/remBoot.Rmd
@@ -3,6 +3,7 @@ title: "remBoot User Guide"
 author: "Anthony Caravaggi (@thonoir; arcaravaggi.co.uk)"
 date: "`r Sys.Date()`"
 output: rmarkdown::html_vignette
+keep_md: true 
 fig_caption: yes
 vignette: >
   %\VignetteIndexEntry{Vignette Title}
@@ -23,11 +24,12 @@ Alternatively, generate relevant simulated data.
 ```{r include = TRUE}
 
 remDat <- data.frame(cbind(site <- sample(1:4, 200, replace=TRUE),
+                           cam <- sample(1:20, 200, replace = TRUE),
                            count <- sample(1:3, 200, replace = TRUE),
                            dist <- runif(200, min=0, max=0.006),
                            theta <- runif(200, min=0, max=0.6)))
 
-colnames(remDat) <- c("site", "count", "dist", "theta")
+colnames(remDat) <- c("site","cam", "count", "dist", "theta")
 
 ```
 
@@ -36,8 +38,8 @@ Now we'll check the first few rows of data, check the sample size for each site,
 
 ```{r include = TRUE}
 head(hDat) 
-tapply(hDat[,4], hDat[,1], length) 
-sum(tapply(hDat[,4], hDat[,1], length)) == nrow(hDat) 
+tapply(hDat[,3], hDat[,1], length) 
+sum(tapply(hDat[,3], hDat[,1], length)) == nrow(hDat) 
 ```
 
 We'll use two functions to split the data according to survey site (__split_dat__) and store the number of rows. These data will be used for iterative resampling during bootstrapping.
