### 2017-02-09 7019c10
Updated re: NAMESPACE
Prevented export of several functions. rem, remBoot and split_dat are the only visible functions.
