+++ R/remBoot.R
@@ -33,21 +33,21 @@
 
 
 remBoot <- function(.df, tm, v, nboots, error_stat){
-    ## Written by Kevin Keenan 2013 and Anthony Caravaggi 2016
-    ## distributed under GPL3
-    grpDat <- split_dat(.df)
-    remsD <- lapply(grpDat, boot_sd)
-    res <- list()
-    if(is.element("sd", error_stat)){
-      resSD <- lapply(remsD, sd)
-      names(resSD) <- levels(.df[[1]])
-      res$sd <- c(res, resSD)
-    }
-    if(is.element("ci", error_stat)){
-      resCI <- t(sapply(remsD, quantile, probs = c(0.025, 0.975)))
-      colnames(resCI) <- c("Lower", "Upper")
-      rownames(resCI) <- levels(.df[[1]])
-      res$ci <- resCI
-    }
-    return(res)
+  ## Written by Kevin Keenan 2013 and Anthony Caravaggi 2016
+  ## distributed under GPL3
+  grpDat <- split_dat(.df)
+  remsD <- lapply(grpDat, boot_sd)
+  res <- list()
+  if(is.element("sd", error_stat)){
+    resSD <- lapply(remsD, sd)
+    names(resSD) <- levels(.df[[1]])
+    res$sd <- c(res, resSD)
+  }
+  if(is.element("ci", error_stat)){
+    resCI <- t(sapply(remsD, quantile, probs = c(0.025, 0.975)))
+    colnames(resCI) <- c("Lower", "Upper")
+    rownames(resCI) <- levels(.df[[1]])
+    res$ci <- resCI
+  }
+  return(res)
 }
