+++ R/boot_sd.R
@@ -2,14 +2,14 @@
 #' 
 #' Resamples camera trap data to calculate variance for Random Encounter Model density estimates. Note that tm and v must be the same across all sites. If this is not the case, split your data before bootstrapping and run the function on each sub-group. Called by remsD. Calls bsD.
 #' @param x = bsD resampled data
-#' @export
 #' @examples 
-#' boot_sd <-  function(x){
-#' d <- replicate(nboots, bsD(x, tm, v)) 
-#' return(d)
-#' }
-#' 
+#' ## Define the number of bootstrapping iterations and apply boot_sd to the data:
+#' nboots <- 1000#' 
 #' remsD <- lapply(grpDat, boot_sd)
+#' 
+#' ## Calculate variance (Standard Deviation)
+#' remsSD <- lapply(remsD, sd)
+#' remsSD
 
 
 boot_sd <-  function(x){
