+++ R/res_dat.R
@@ -3,7 +3,6 @@
 #' Use known row number to tell function how many data to resample in each iteration
 #' 
 #' @param .df = the data frame
-#' @export
 #' @examples 
 #' 
 #' res_dat <- function(.df) {
@@ -14,5 +13,5 @@
 
 
 res_dat <- function(.df) {
-  tapply(.df[,3], .df[,1], length)
+  tapply(.df[,4], .df[,1], length)
 }
