### 2017-02-09 522ad14
NAMESPACE limited export
Hidden several non-user-facing functions.
