+++ NAMESPACE
@@ -1,9 +1,5 @@
 # Generated by roxygen2: do not edit by hand
 
-export(boot_sd)
-export(bsD)
-export(grpDat_func)
 export(rem)
 export(remBoot)
-export(res_dat)
 export(split_dat)
