### 2017-02-09 fe04c96
Removed from Help file
Non user-facing functions. Script removed to omit from Help file.
