+++ man/bsD.Rd
@@ -1,39 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/bsd.R
-\name{bsD}
-\alias{bsD}
-\title{Required for remsD bootstrapping}
-\usage{
-bsD(dat, tm, v)
-}
-\arguments{
-\item{dat}{= the data frame}
-
-\item{tm}{= camera hours}
-
-\item{v}{= distance travelled in 24 hours (km)}
-}
-\description{
-Resamples camera trap data to calculate variance for Random Encounter Model density estimates. Called by remsD.
-}
-\examples{
-
-bsD <- function(dat, tm, v){
-bsDat <- dat[sample(1:nrow(dat), size = nrow(dat), replace = TRUE), ]
-dOut <- rem(bsDat, tm, v)
-return(dOut)
-}
-
-
-remsD <- lapply(grpDat, function(x){
- d <- replicate(nboots, bsD(dat, tm, v)) 
- return(d)
-})
-}
-\keyword{bootstrap}
-\keyword{density,}
-\keyword{encounter}
-\keyword{model,}
-\keyword{population,}
-\keyword{random}
-
