### 2017-02-09 bbc9af0
Delete grpDat_func.Rd
