+++ man/grpDat_func.Rd
@@ -1,27 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/grpDat_func.R
-\name{grpDat_func}
-\alias{grpDat_func}
-\title{Required for split_dat}
-\usage{
-grpDat_func(df, x)
-}
-\arguments{
-\item{df}{= the data frame}
-
-\item{x}{= group (site), derived from df}
-}
-\description{
-Separate data into groups prior to bootstrapping. Called by split_dat.
-}
-\examples{
-
-grpDat_func <- function(df, x){
-return(df[which(df[,1] == x), ])
-}
-
-split_dat <- function(.df) {
-  lapply(unique(.df[[1]]), grpDat_func, df = .df)
-}
-}
-
