### 2017-02-09 54680bd
Delete res_dat.Rd
