+++ man/remBoot.Rd
@@ -7,24 +7,40 @@
 remBoot(.df, tm, v, nboots, error_stat)
 }
 \arguments{
-\item{.df}{= the data frame}
+\item{.df}{(data frame)}
 
-\item{tm}{= camera hours}
+\item{tm}{(numeric) The total number of hours all cameras were left in-situ at a focal site}
 
-\item{v}{= distance travelled in 24 hours (km)}
+\item{v}{(numeric) The distance travelled by the focal species in 24 hours, in kilometres}
 
-\item{nboots}{= number of bootstrap iterations}
+\item{nboots}{(numeric) The number of bootstrap iterations}
 
-\item{error_stat}{= confidence intervals (ci) or standard deviation (sd)}
+\item{error_stat}{(text) Variance, given as confidence intervals (ci) or standard deviation (sd)}
 }
 \description{
 This function allows the automated calculation of variance for animal densities estimates derived from Random Encounter Modelling.
 }
 \details{
+The function assumes that the first 4 columns of the dataset contain: 
+1) An identifying number for each survey location (e.g. 1, 2, 3)
+2) The number of individuals of the focal species observed in each capture
+3) The radial distance to the detected animal in each capture, given in metres
+4) The angle of detection based on the location of the detected animal in each capture, given in radians
+
 For an example of how to use remBoot, click (\href{http://htmlpreview.github.io/?https://github.com/arcaravaggi/remBoot/blob/master/vignettes/remBoot.html}{here}).
 }
 \examples{
-output <- remBoot(remDat, tm, v, nboots, error_stat = c("sd"))
+## Define tm, v and nboots and pass the values to the function:
+
+tm <- 1800
+v <- 0.89
+nboots <- 1000
+
+output <- remBoot(.df, tm, v, nboots, error_stat = c("sd"))
+
+## Alternatively, pass the values to the function, directly:
+
+output <- remBoot(hDat, tm = 1800, v = 0.89, nboots, error_stat = c("ci"))
 }
 \keyword{bootstrap}
 \keyword{density,}
