### 2017-02-09 83be9a0
Rectified error
Removed @export entirely - functions stopped working. Whoops!
