+++ R/grpDat_func.R
@@ -1,3 +1,5 @@
+#' @export
+
 grpDat_func <- function(df, x){
   return(df[which(df[,1] == x), ])
 }
