+++ R/boot_sd.R
@@ -2,6 +2,7 @@
 #' 
 #' Resamples camera trap data to calculate variance for Random Encounter Model density estimates. Note that tm and v must be the same across all sites. If this is not the case, split your data before bootstrapping and run the function on each sub-group. Called by remsD. Calls bsD.
 #' @param x = bsD resampled data
+#' @export
 #' @examples 
 #' ## Define the number of bootstrapping iterations and apply boot_sd to the data:
 #' nboots <- 1000#' 
