+++ man/boot_sd.Rd
@@ -14,10 +14,8 @@ Resamples camera trap data to calculate variance for Random Encounter Model dens
 }
 \examples{
 ## Define the number of bootstrapping iterations and apply boot_sd to the data:
-nboots <- 1000#' 
+nboots <- 1000
 remsD <- lapply(grpDat, boot_sd)
-
-## Calculate variance (Standard Deviation)
 remsSD <- lapply(remsD, sd)
 remsSD
 }
