### 2017-02-09 c22749d
Run into an error
The local package runs as anticipated. GitHub installs run into an error. I'm uploading local files in an attempt to rectify this.
