### 2017-02-09 #176 @arcaravaggi
I've dealt with a number of the corrections and will be applying myself to the others (functionality clarification, tests, sanity checks) presently. Please let me know whether the corrections entered thus far are insufficient and/or inappropriate.

- Release 0.0.0.9000 created.
- Installation instructions added to the README.
- Moved package introduction text from vignette to the README.
- Included a brief, un-annotated example in the README.
- Corrected function documentation in `remboot`, described data requirements and expanded the example contained within.
- Corrected function documentation in `rem`, described data requirements and expanded the example contained within.
- Removed column 2 (legacy) from data and edited relevant files. This wasn't requested, but makes things a little neater as column 2 wasn't used.
- Edited Help docs for all user-facing functions.
- Added guidelines to README re: contributions and the reporting of errors.
- Corrected NAMESPACE to only export `split_dat`, `rem` and `remboot`.
- Added instruction re: _nboots_ to `remBoot` and examples.
- `grpDat_func` isn't user-facing and has been hidden.
- `remSD` is merely a vector used to store bootstrapping results: remsD <- lapply(grpDat, boot_sd)
