### 2017-02-09 #176 @arcaravaggi
Re: the external link to the vignette in `rem`. As I understand it, the html file needs to be parsed by http://htmlpreview.github.io/... for the output to be presented, as opposed to code. Is this not the case?
