### 2017-02-09 #176 @amoeba
Thanks for looking at these. I'll take a look over them soon, at least by this weekend.

Regarding vignettes, I'm not familiar enough with them off-hand but usually I can run `vignette(package="remBoot")` to get a list of the vignettes provided in `remBoot` and then, if I find a vignette called 'remBoot' in that list, I can the run `vignette('remBoot', package='remBoot')` to view a nicely-rendered vignette inside R (locally). I can look into this as it's something I should learn anyway.
