### 2017-02-13 #176 @amoeba
Thanks for making all of those changes. You'll notice I've checked off a number of the list items. Some minor details:

- Given the number of changes to the code, I suggest this submission get a new release of the package with the changes before finalizing this submission.

Documentation:

- The updated README is a good improvement. I'll sign off on this as-is but I suggest using code fences to provide rich markup for the code and also fix two broken references (Rowcliffe citation and Fig 2 reference)
- On the help doc for `rem`, the names in the function signature and the argument documentation are out of alignment (dat vs. .d.f)

Minor, not-crucial-for-review issues:

- I'm still a little perplexed about the `nboots` argument in `remBoot`. The way you've coded it, the argument `nboots` is never used and the value that is used by `boot_sd` is taken from the calling environment. I would expect this to be confusing to most users. Would you consider changing the way this works? If not, I suggest removing the `nboots` argument from the function and its docs and documentation the behavior elsewhere.
