+++ README.md
@@ -1,20 +1,19 @@
 # remBoot
 
-remBoot is an implementation of the Random Encounter Model (REM) developed by Rowcliffe _et al._ (2008). The REM is based on brownian motion and allows the estimation of animal densities for a given survey area. The REM is particularly useful in that species do not need to be marked or exhibit individual markings for estimates to be calculated.   
-
-remBoot is the first package to implement REM calculations in R. The package also contains functions which allow the calculation of variance (standard deviation [SD] and/or 95% confidence intervals [CI]; Fig. 2). These calculations are computationally inexpensive and can be applied to datasets of considerable size. Densities and associated variances can be calculated for one or more sites concurrently, streamlining the analytical process.  
+remBoot is an implementation of the Random Encounter Model (REM) developed by Rowcliffe _et al._ (2008). The REM is based on brownian motion and allows the estimation of animal densities for a given survey area. The REM is particularly useful in that species do not need to be marked or exhibit individual markings for estimates to be calculated. This package contains a number of functions which allow the user to calculate densities for one or more sites and bootstrap their data to calculate variance.   
 
 # Installation
 
 To install remBoot from the GitHub repo, use the following commands:  
-
+```
 install.packages("githubinstall")  
 library(githubinstall)  
 githubinstall("remBoot")  
-
+```
 
 # Example
 
+```
 data(hDat)  
 grpDat <- split_dat(hDat)  
 tm <- 3600  
@@ -26,7 +25,7 @@ nboots <- 1000
 remsD <- lapply(grpDat, boot_sd)   
 remsSD <- lapply(remsD, sd)  
 remsSD  
-
+```
 A detailed example can be found at http://htmlpreview.github.io/?https://github.com/arcaravaggi/remBoot/blob/master/vignettes/remBoot.html  
 
 # Development
@@ -41,6 +40,9 @@ Please report it to the [issue tracker][issues]. Please provide specific details
 
 [issues]: https://github.com/arcaravaggi/remBoot/issues
 
+# References
+
+Rowcliffe, J.M., Field, J., Turvey, S.T. and Carbone, C. (2008). Estimating animal density using camera traps without the need for individual recognition. J Appl Ecol. 45, 1228 – 1236. [DOI: 10.1111/j.1365-2664.2008.01473.x](http://onlinelibrary.wiley.com/doi/10.1111/j.1365-2664.2008.01473.x/abstract)
 
 # License
 
