### 2017-02-16 ce58c48
Minor edits to paper.md
