+++ R/remBoot.R
@@ -28,7 +28,7 @@
 #' 
 #' ## Alternatively, pass the values to the function, directly:
 #' 
-#' output <- remBoot(hDat, tm = 1800, v = 0.89, nboots, error_stat = c("ci"))
+#' output <- remBoot(hDat, tm = 1800, v = 0.89, nboots = 1000, error_stat = c("ci"))
 
 
 
