### 2017-02-16 #176 @arcaravaggi
Thanks for the additional comments.

- Would you be happy for me to create a new release when we have completed the review process (prior to finalising)? If not, I can do it sooner, no problem.
- I've added code fences to README and fixed the Rowcliffe citation. I'm not sure what's wrong with the Fig. 2 reference, though.
- Re: `nboots` in `remBoot`, the parameter can either be defined in the calling environment or passed to `remBoot', as with _tm_ and _v_. I have added a little text to the `remBoot` example, demonstrating this.

Regarding functionality (in your first response), users are able to calculate densities in two ways with `rem`: 1) by passing the dataframe to the function directly, thereby calculating densities across all data, or (2) by using grpDat for site-specific densities (assuming multiple sites). This is explained in the vignette.

Looking back at your comments, I think there are 3 issues outstanding (not including any carry-over from my responses, above):

- Add documentation on handling NAs
- Sanity checking (I'm looking into it of course, but is this essential for the purposes of the review?)
- Automated tests

Please do let me know if I've missed anything.
