### 2017-02-16 #176 @amoeba
> Would you be happy for me to create a new release when we have completed the review process (prior to finalising)? If not, I can do it sooner, no problem.

Yes, doing just one release is ideal.

> I'm not sure what's wrong with the Fig. 2 reference, though.

I expected there to be a figure 2 in the readme which there wasn't.

Your changes and responses are sufficient. Thanks for addressing those. I'm having a hard time actually running `remBoot` right now though:

```
> devtools::load_all(".")
> remBoot(hDat, tm = 1800, v = 0.89, nboots, error_stat = c("ci"))
Error in rem(bsDat, tm, v) : object 'tm' not found
```

I haven't looked at the cause or at recent commits to see what's going on here but I thought I'd bring it to your attention.

Other than that, the automated tests are the only thing I think we need.
