### 2017-02-16 #176 @arcaravaggi
Hrm, that's interesting. I might be away for a few days, so will probably have to come back to this early next week. Sorry about that.

I've not written any automated tests before, so I'll have to do some reading. Do you have any recommendations on sources and/or implementation?
