### 2017-02-16 #176 @amoeba
A useful intro and guide is http://r-pkgs.had.co.nz/tests.html.  A couple of tests for `remBoot` that test the return value for structure and correctness would probably be enough to check off the automated tests checklist item. It would catch the type of issue I received above (if in fact my issue above was not my fault, which it may be).

If you'd like, I can stub out testing and a few basic tests and push them your way.
