### 2017-02-16 #176 @arcaravaggi
That'd be much appreciated, thank you.

What would be the best way to acknowledge your efforts as reviewer, by the way? I usually mention reviewers in the acknowledgements of my papers, but JOSS is a bit different.
