### 2017-02-16 #176 @amoeba
I don't know, actually. Perhaps @karthik or someone at JOSS knows. I don't think there are any guidelines preventing you from putting acknowledgements in your `paper.md` so I'd probably put things there.
