### 2017-02-17 db618b5
Minor edit for functionality
