### 2017-02-17 #176 @arcaravaggi
I've been unable to replicate your error this morning. I've re-downloaded the package and everything works on my end. Could you try reinstalling to see whether the error persists?

A quick note on `nboots` in `remBoot`: The function wasn't accepting the parameter if fed directly, so I've embedded the `boot_sd` function within `remBoot` and that seems to have fixed the issue. `boot_sd` is also still available as a standalone function for use with `rem`.
