### 2017-02-17 #176 @amoeba
Okay, maybe it was my fault. I'll double-check and get back to you. Thanks for the fix on the `nboots` param.
