### 2017-02-18 a89eb35
Minor changes to README
