+++ DESCRIPTION
@@ -15,5 +15,7 @@ Encoding: UTF-8
 LazyData: true
 RoxygenNote: 5.0.1
 Suggests: knitr,
-    rmarkdown
+    rmarkdown,
+    testthat,
+	ggplot2
 VignetteBuilder: knitr
