### 2017-02-21 #176 @amoeba
You had previously said:

> Corrected NAMESPACE to only export split_dat, rem and remboot

I noticed the roxygen tags didn't match with this so I removed extra @export tags. Then I noticed `boot_sd` is used in the vignette so I exported `boot_sd`.

I tweaked this stuff and also ran and re-ran `devtools::check` and `devtools::document` until things ran smoothly (fixing issues along the way). See my PR. Unfortunately, it looks like RStudio clobbered your line endings, making the git diff less than useful.

Hopefully my set of documentation changes are agreeable and my example tests are useful. I'll review the checklist once some or all of my changes are merged.
