### 2017-02-21 #176 @arcaravaggi
I left the export tags on the other functions as they didn't work locally without them, for some reason. If they work now, that's great, thanks.

RStudio really has done a number on my lines. I'll comb through and try to identify the changes.
