### 2017-02-21 #176 @amoeba
Yeah I can re-PR with a non-borked version if you like. Are you on Windows? I have my RStudio set up to save UNIX-style line endings but it appears I can change them to Windows-style.
