### 2017-02-21 arcaravaggi/remBoot#17 @amoeba
R CMD Check and devtools::document() issues
I'm not having any luck with having RStudio not clobber my edits into useless diffs so I'll make my comments here. My process for finding these issues is run and re-run `devtools::document` and `devtools::check` until the issues go away.

- Vignette doesn't load `remBoot` at the start and needs to
- Remove @export tags or document these:

```
* checking for missing documentation entries ... WARNING
Undocumented code objects:
  ‘grpDat_func’ ‘res_dat’
All user-level objects in a package should have documentation entries.
See chapter ‘Writing R documentation files’ in the ‘Writing R
Extensions’ manual.
```

- Fix a typo in the docs for `rem`

```
* checking Rd \usage sections ... WARNING
Undocumented arguments in documentation object 'rem'
  ‘dat’
Documented arguments not in \usage in documentation object 'rem':
  ‘.df’

Functions with \usage entries need to have the appropriate \alias
entries, and all their arguments documented.
The \usage entries must correspond to syntactically valid R code.
See chapter ‘Writing R documentation files’ in the ‘Writing R
Extensions’ manual.
```
- Examples don't run:

```
* checking examples ... ERROR
Running examples in ‘remBoot-Ex.R’ failed
The error most likely occurred in:

> base::assign(".ptime", proc.time(), pos = "CheckExEnv")
> ### Name: boot_sd
> ### Title: Required for remsD bootstrapping
> ### Aliases: boot_sd
>
> ### ** Examples
>
> ## Define the number of bootstrapping iterations and apply boot_sd to the data:
> nboots <- 1000
> remsD <- lapply(grpDat, boot_sd)
Error in lapply(grpDat, boot_sd) : object 'grpDat' not found
Execution halted
```

There may be more issues after that is fixed.

- Missing import statements:

```
checking R code for possible problems ... NOTE
boot_sd: no visible binding for global variable ‘nboots’
boot_sd: no visible binding for global variable ‘tm’
boot_sd: no visible binding for global variable ‘v’
remBoot: no visible binding for global variable ‘sd’
remBoot: no visible binding for global variable ‘quantile’
  nboots quantile sd tm v
Consider adding
  importFrom("stats", "quantile", "sd")
to your NAMESPACE file.
```
