### 2017-02-21 #176 @amoeba
I was unable to get my edits with RStudio to produce meaningful diffs so I made an issue and pasted what I was finding: https://github.com/arcaravaggi/remBoot/issues/17. Could you please take a look and look into fixing those issues?
