### 2017-02-21 arcaravaggi/remBoot#17 @arcaravaggi
I'll be uploading fixes as a batch, rather than individually.

- Fixed the vignette `remboot` issue.
- Corrected the .df-dat issue in `rem`
- Added documentation in `boot_sd`. However, the note 'no visible binding for global variable [variable name] aren't concerns as they're defined earlier and passed into the function (_tm_, _v_ & _nboots_)
- Removed @export from `grpDat_func` and `res_dat`. Also from `bsD` (it appears to work...)
- NAMESPACE is overwritten every time the package is rebuilt during checking. I've added importFrom("stats", "quantile", "sd") manually.

I'm not getting any remaining issues from the checks.
