### 2017-02-21 96e9cbd
Corrections & amendments based on devtools checks
