+++ R/res_dat.R
@@ -1,5 +1,3 @@
-#' @export
-
 res_dat <- function(.df) {
   tapply(.df[,4], .df[,1], length)
 }
