+++ R/grpDat_func.R
@@ -1,5 +1,3 @@
-#' @export
-
 grpDat_func <- function(df, x){
   return(df[which(df[,1] == x), ])
 }
