+++ vignettes/remBoot.knit.md
@@ -1,7 +1,7 @@
 ---
 title: "remBoot User Guide"
 author: "Anthony Caravaggi (@thonoir; arcaravaggi.co.uk)"
-date: "`r Sys.Date()`"
+date: "2017-02-21"
 output: rmarkdown::html_vignette
 keep_md: true 
 fig_caption: yes
@@ -13,7 +13,8 @@ vignette: >
 
 Read the packaged data file. These data are from a study of hares in Northern Ireland, published as Caravaggi _et al._ (2016). The first 5 columns of any data set must follow the same format in terms of column contents.
 
-```{r include = TRUE}
+
+```r
 library(remBoot)
 
 data(hDat)
@@ -21,8 +22,8 @@ data(hDat)
 
 Alternatively, generate relevant simulated data. 
 
-```{r include = TRUE}
 
+```r
 remDat <- data.frame(cbind(site <- sample(1:4, 200, replace=TRUE),
                            cam <- sample(1:20, 200, replace = TRUE),
                            count <- sample(1:3, 200, replace = TRUE),
@@ -30,122 +31,281 @@ remDat <- data.frame(cbind(site <- sample(1:4, 200, replace=TRUE),
                            theta <- runif(200, min=0, max=0.6)))
 
 colnames(remDat) <- c("site","cam", "count", "dist", "theta")
-
 ```
 
 
 Now we'll check the first few rows of data, check the sample size for each site, and check data consistency.
 
-```{r include = TRUE}
+
+```r
 head(hDat) 
+```
+
+```
+##   site count  dist theta
+## 1    1     1 0.003  0.67
+## 2    1     1 0.003  0.67
+## 3    1     1 0.003  0.67
+## 4    1     1 0.004  0.39
+## 5    1     1 0.003  0.18
+## 6    1     1 0.005  0.10
+```
+
+```r
 tapply(hDat[,3], hDat[,1], length) 
+```
+
+```
+##  1  2  3  4  5 
+## 36 16 20 19 27
+```
+
+```r
 sum(tapply(hDat[,3], hDat[,1], length)) == nrow(hDat) 
+```
+
+```
+## [1] TRUE
 ```
 
 We'll use two functions to split the data according to survey site (__split_dat__) and store the number of rows. These data will be used for iterative resampling during bootstrapping.
 
-```{r include = TRUE}
 
+```r
 grpDat <- split_dat(hDat)
-
 ```
 
 Check that the code has done what we wanted
 
-```{r include = TRUE}
-lapply(grpDat, head)
 
+```r
+lapply(grpDat, head)
+```
+
+```
+## [[1]]
+##   site count  dist theta
+## 1    1     1 0.003  0.67
+## 2    1     1 0.003  0.67
+## 3    1     1 0.003  0.67
+## 4    1     1 0.004  0.39
+## 5    1     1 0.003  0.18
+## 6    1     1 0.005  0.10
+## 
+## [[2]]
+##    site count  dist theta
+## 37    2     1 0.001  1.37
+## 38    2     1 0.006  0.06
+## 39    2     1 0.003  0.30
+## 40    2     1 0.002  0.18
+## 41    2     1 0.003  0.17
+## 42    2     1 0.005  0.51
+## 
+## [[3]]
+##    site count  dist theta
+## 53    3     1 0.007  0.29
+## 54    3     1 0.006  0.24
+## 55    3     1 0.006  0.12
+## 56    3     1 0.005  0.19
+## 57    3     1 0.002  0.31
+## 58    3     1 0.004  0.16
+## 
+## [[4]]
+##    site count  dist theta
+## 73    4     1 0.003  0.28
+## 74    4     1 0.007  0.18
[…]
