### 2017-02-21 arcaravaggi/remBoot#17 @amoeba
Great, thanks. Add the @importFrom to the roxygen block for the function to get it to show up in NAMESPACE
