+++ NAMESPACE
@@ -4,4 +4,5 @@ export(boot_sd)
 export(rem)
 export(remBoot)
 export(split_dat)
-importFrom("stats", "quantile", "sd")
+importFrom(stats,quantile)
+importFrom(stats,sd)
