### 2017-02-21 ee9b7a2
Added importFrom
stats quantile sd
