### 2017-02-21 arcaravaggi/remBoot#17 @arcaravaggi
Done. Thanks for your patience. 3 notes remain, which aren't of concern:

```checking R code for possible problems ... NOTE
boot_sd: no visible binding for global variable 'nboots'
boot_sd: no visible binding for global variable 'tm'
boot_sd: no visible binding for global variable 'v'
  nboots tm v```
