### 2017-02-21 arcaravaggi/remBoot#17 @amoeba
Excellent, thanks for addressing those. I'll take a look and see if we can finish this out today.
