### 2017-02-21 arcaravaggi/remBoot#17 @amoeba
On a fresh pull from master on this repo, I still get an error during `devtools::check`

```
* checking examples ... ERROR
Running examples in ‘remBoot-Ex.R’ failed
The error most likely occurred in:

> base::assign(".ptime", proc.time(), pos = "CheckExEnv")
> ### Name: bsD
> ### Title: Required for remsD bootstrapping
> ### Aliases: bsD
> ### Keywords: bootstrap density, encounter model, population, random
>
> ### ** Examples
>
>
> bsD <- function(dat, tm, v){
+ bsDat <- dat[sample(1:nrow(dat), size = nrow(dat), replace = TRUE), ]
+ dOut <- rem(bsDat, tm, v)
+ return(dOut)
+ }
>
>
> remsD <- lapply(grpDat, function(x){
+  d <- replicate(nboots, bsD(dat, tm, v))
+  return(d)
+ })
Error in lapply(grpDat, function(x) { : object 'grpDat' not found
Execution halted
```
