### 2017-02-21 ab2b8bf
Delete bsd.R
