### 2017-02-21 arcaravaggi/remBoot#17 @arcaravaggi
My last push didn't remove the bsd.R file (supplanted by the more appropriately named bsD.R) for some reason. I've removed it, done a fresh pull, and got a clean check (apart from those 3 notes).
