### 2017-02-21 arcaravaggi/remBoot#17 @arcaravaggi
Great.  Are we good to close this issue?
