### 2017-02-21 #176 @amoeba
@arcaravaggi and I worked from https://github.com/arcaravaggi/remBoot/issues/17 to address my remaining issues. I have now checked the automated tests list item as I wrote up a minimal set of regression tests. This package isn't particularly involved so a much larger test suite might not be strictly necessary.

I change my review to Accept. We need a new release of the R package to accompany the JOSS review at which point I think everything is done here.
