### 2017-02-21 #176 @arcaravaggi
I have created a new release to accompany the JOSS review: https://github.com/arcaravaggi/remBoot/releases/tag/v0.1.0
