### 2017-02-23 #176 @arcaravaggi
I have created a new figshare archive as required by JOSS: https://doi.org/10.6084/m9.figshare.4684894

Once the JOSS article is published and a DOI has been assigned, I will update the package's README with citation information, include a CITATION file, and add the JOSS article DOI to the figshare archive in the form of a comment.

It is likely that there will be developments in Random Encounter Modelling from camera trap data in the future. @arfon, are there any limits on continued development of packages which have been accepted for publication in JOSS?
