### 2017-02-28 #176 @arfon
I think next steps are on me now to accept this submission.

>  @arfon, are there any limits on continued development of packages which have been accepted for publication in JOSS?

@arcaravaggi - nope. This is totally fine 😁
