### 2017-02-28 #176 @arfon
@arcaravaggi - could you please move your references into a `paper.bib` file and cite them inline from the `paper.md` markdown file? We compile the paper using pandoc which handles the production of the PDF etc.
