### 2017-02-28 #176 @arfon
@amoeba many thanks for reviewing this submission ✨

@arcaravaggi - your paper is now accepted into JOSS and your paper DOI is http://dx.doi.org/10.21105/joss.00176 :zap: 🚀 💥
