### 2017-03-18 #176 @karthik
Sorry I was away on vacation when the last few discussions happened and there was no way to leave a vacation notice.

@amoeba It is possible to add @arcaravaggi as a reviewer in your package DESCRIPTION using that role. This is still possible after acceptance.  You can do it in your next CRAN update.
